Syncordian.Byzantine (syncordian v0.1.0)

This module is responsible for the byzantine operations aiming to provide the features to support the byzantine fault tolerance in the Syncordian implementation.

Functions

check_signature(left_parent, to_check_line, right_parent)

@spec check_signature(
  left_parent :: Syncordian.Types.line(),
  to_check_line :: Syncordian.Types.line(),
  right_parent :: Syncordian.Types.line()
) :: boolean()

create_signature(left_parent_signature, new_content, right_parent_signature, peer_id)

@spec create_signature(
  left_parent_signature :: Syncordian.Types.content(),
  new_content :: Syncordian.Types.content(),
  right_parent_signature :: Syncordian.Types.content(),
  peer_id :: Syncordian.Types.peer_id()
) :: Syncordian.Types.signature()

This function creates a signature based on the left parent content, the new content, the right parent content and the peer id by concatenating them and hashing the result using the sha256 algorithm. It returns the hash encoded in base16.