Syncordian.Line (syncordian v0.1.0)

This module is responsible for the line structure and the line operations provides the following functions:

get_line_id(line) : returns the line id

get_content(line) : returns the content of the line

get_signature(line) : returns the signature of the line

get_peer_id(line) : returns the peer id of the line

create_infimum_line(peer_id) : creates the infimum line for the given peer id document

create_supremum_line(peer_id) : creates the supremum line for the given peer id document

create_line_between_two_lines(content, left_parent, right_parent) : creates a new line between two lines

Functions

compare_lines(line1, line2)

@spec compare_lines(
  line1 :: Syncordian.Types.line(),
  line2 :: Syncordian.Types.line()
) :: 0 | 1 | -1

Given two lines this function defines the order between them, mainly based on the line_id field of the lines and the usual number comparison. It returns:

- 1 if the first line is greater than the second line
- 0 if the first line is equal to the second line
- -1 if the first line is less than the second line
TODO: When is the case that the distance between the lines is 1 is important review
this, I think that in the case the 0 option should be returned (?)

create_line_between_two_lines(content, left_parent, right_parent)

@spec create_line_between_two_lines(
  content :: Syncordian.Types.content(),
  left_parent :: Syncordian.Types.line(),
  right_parent :: Syncordian.Types.line()
) :: Syncordian.Types.line()

Given two lines, left_parent and right_parent, this function creates a new line between them taken into account the parent's ids to calculate the new line id and the parent's content to create the line signature. There are two cases to consider:

- If there is 'room' between the parent's ids, then the new line id is calculated
  randomly between the parent's ids and the signature is created using the parent's
  content and the peer id.
- If there is no 'room' between the parent's ids, then...