Syncordian.Line_Object (syncordian v0.1.0)

Functions

create_infimum_line(peer_id)

@spec create_infimum_line(Syncordian.Types.peer_id()) :: Syncordian.Types.line()

This function creates the infimum line for the given peer id that is the absolute first line within peer's document

create_supremum_line(peer_id)

@spec create_supremum_line(Syncordian.Types.peer_id()) :: Syncordian.Types.line()

This function creates the supremum line for the given peer id that is the absolute last line within peer's document

get_content(line)

@spec get_content(Syncordian.Types.line()) :: Syncordian.Types.content()

This function is a getter for the content field of a line record

get_line_id(line)

@spec get_line_id(Syncordian.Types.line()) :: Syncordian.Types.line_id()

This function is a getter for the line_id field of a line record

get_line_insertion_attempts(line)

get_peer_id(line)

@spec get_peer_id(Syncordian.Types.line()) :: Syncordian.Types.peer_id()

This function is a getter for the peer_id field of a line record

get_signature(line)

@spec get_signature(Syncordian.Types.line()) :: Syncordian.Types.signature()

This function is a getter for the signature field of a line record

get_status(line)

@spec get_status(Syncordian.Types.line()) :: boolean()

This function is a getter for the deleted field of a line record, this field is true when the line was marked as deleted false otherwise

line(args \\ [])
(macro)

line(record, args)
(macro)

set_line_status(line, new_status)

tick_line_insertion_attempts(line)