Syncordian.Site (syncordian v0.1.0)

This module is responsible for the site structure and the site operations provides the following functions:

start(peer_id) : starts a site with the given peer_id

insert(pid,content,index) : inserts a content at the given index

info(pid) : prints the document of the site

raw_print(pid) : prints the document of the site without the site structure

Functions

delete_line(pid, index_position)

insert(pid, content, index_position)

@spec insert(pid(), String.t(), integer()) :: any()

This function inserts a content at the given index and a pid by sending a message to the loop site function. The messages uses the following format:

loop(site)

@spec loop(Syncordian.Types.site()) :: any()

This function is the main loop of the site, it receives messages and calls the appropriate functions to handle them.

raw_print(pid)

@spec raw_print(pid()) :: any()

This function prints the whole document as a list of lists by sending a message to the loop site function with the atom :print.

site(args \\ [])
(macro)

site(record, args)
(macro)

start(peer_id)

@spec start(Syncordian.Types.peer_id()) :: pid()

This function starts a site with the given peer_id and registers it in the global registry. The returned content is the pid of the site. The pid is the corresponding content of the pid of the spawned process.