Syncordian.Types (syncordian v0.1.0)

This module provides the types used in the Syncordian implementation

Types

clock()

@type clock() :: integer()

content()

@type content() :: String.t()

document()

@type document() :: [line()]

Type that represents a Syncordian document, that is a list of lines, each line has an
unique position identifier and a content

line()

@type line() :: %{
  line_id: line_id(),
  content: content(),
  signature: signature(),
  peer_id: peer_id(),
  status: boolean(),
  insertion_attempts: integer()
}

Type that represents a line of a Syncordian document

line_id()

@type line_id() :: float()

Type that represents the position identifier of a Syncordian document line 

peer_id()

@type peer_id() :: integer()

Type that represents the peer id

signature()

@type signature() :: String.t()

site()

@type site() :: %{id: peer_id(), clock: clock(), document: document(), pid: pid()}

Type that represents the individual replica of the Syncordian document for each peer